Create README.md
You are a Data Analyst working for an e-commerce company named X. You are tasked with preparing a presentation to present an overview of the business and operations of the company up to now. for Sales and Operations Managers. The presentation should at a minimum include the following information: business overview, customer satisfaction, and 2 to 3 areas of recommendation (areas) where the company can improve.

Some additional information for the case study: Since there is only data up to 2018, we will assume that it is currently September 2018 (data after September 2018 you can ignore) The company is based in the US, however incorporated in Brazil (that's why some information is written in Portuguese)